requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
chromadb==0.4.22
openai==1.3.7
anthropic==0.8.1
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-docx
